common web, refactor route finding/handling
* less duplication between alertmanager api and rest of the web handlers
* no need to replicate the context key
